+++ cmd/frontend/graphqlbackend/search.go
@@ -245,7 +245,7 @@ func (r *searchResolver) resolveRepositories(ctx context.Context, effectiveRepoF
 	archivedStr, _ := r.query.StringValue(query.FieldArchived)
 	archived := parseYesNoOnly(archivedStr)
 
-	commitSince, _ := r.query.StringValue(query.FieldRepoHasCommitSince)
+	commitSince, _ := r.query.StringValue(query.FieldCommitSince)
 
 	tr.LazyPrintf("resolveRepositories - start")
 	repoRevs, missingRepoRevs, repoResults, overLimit, err = resolveRepositories(ctx, resolveRepoOp{