+++ cmd/frontend/internal/pkg/search/query/searchquery.go
@@ -9,17 +9,17 @@ import (
 
 // All field names.
 const (
-	FieldDefault            = ""
-	FieldCase               = "case"
-	FieldRepo               = "repo"
-	FieldRepoGroup          = "repogroup"
-	FieldFile               = "file"
-	FieldFork               = "fork"
-	FieldArchived           = "archived"
-	FieldLang               = "lang"
-	FieldType               = "type"
-	FieldRepoHasFile        = "repohasfile"
-	FieldRepoHasCommitSince = "repohascommitsince"
+	FieldDefault     = ""
+	FieldCase        = "case"
+	FieldRepo        = "repo"
+	FieldRepoGroup   = "repogroup"
+	FieldFile        = "file"
+	FieldFork        = "fork"
+	FieldArchived    = "archived"
+	FieldLang        = "lang"
+	FieldType        = "type"
+	FieldRepoHasFile = "repohasfile"
+	FieldCommitSince = "commitsince"
 
 	// For diff and commit search only:
 	FieldBefore    = "before"
@@ -51,8 +51,8 @@ var (
 			FieldLang:      {Literal: types.StringType, Quoted: types.StringType, Negatable: true},
 			FieldType:      stringFieldType,
 
-			FieldRepoHasFile:        regexpNegatableFieldType,
-			FieldRepoHasCommitSince: {Literal: types.StringType, Quoted: types.StringType, Singular: true},
+			FieldRepoHasFile: regexpNegatableFieldType,
+			FieldCommitSince: {Literal: types.StringType, Quoted: types.StringType, Singular: true},
 
 			FieldBefore:    stringFieldType,
 			FieldAfter:     stringFieldType,