+++ web/src/nav/NavLinks.test.tsx
@@ -61,6 +61,8 @@ describe('NavLinks', () => {
         keybindings: KEYBINDINGS,
         settingsCascade: SETTINGS_CASCADE,
         history,
+        isSourcegraphDotCom: false,
+        showStatusIndicator: false,
     }
 
     // The 3 main props that affect the desired contents of NavLinks are whether the user is signed